backend/requirements.txt
Django
django-cors-headers
djangorestframework
python-dotenv
google-generativeai